Implement volume copy
Container Hosts environments do not support bind mounts. This sucks as it
means we'll have to manually copy the /local and /secrets alloc data to
a Docker volume before starting the task. We use the same code as
the 'docker cp' command for this to achieve maximum compatibility. Using
large artefacts in your jobs will have an impact on the startup time and
storage usage. Each allocation will effectively have 2 copies of the
artefacts, one in the Nomad client container and a second one in the
dedicated alloc volume. Ironically, we can work around this "hardening"
step because the Docker API endpoint is left wide open! LOL!